added tagging functionality
each thought can have multiple tags associated with it by specifiying
them in a comma-delimited string in detail view.  A new TagsScreen lists
all available tags in the app. Clicking on one will result in a new
route (thoughts/tag/:tag) being navigated to.  This will render the
thoughts screen but only list those thoughts that have been tagged with
the tag specified in the route.