reduced injector usage to locations where it adds value
view creation using the injector is now limited to the content region in
Layout.js.  This eliminates having to pass in the entire context object
to every view whilst still not having to customise what dependencies are
passed depending on the view.  ThoughtDetail and ThoughtsList (both sub
views) didn't really need the injector so I removed it for these views